+++ pyemma/coordinates/clustering/include/regspace.h
@@ -5,8 +5,20 @@
 #ifndef PYEMMA_REGSPACE_H
 #define PYEMMA_REGSPACE_H
 
+#if defined(USE_OPENMP)
+#include <omp.h>
+#endif
 
 #include <Clustering.h>
+
+
+class MaxCentersReachedException : public std::exception {
+public:
+    explicit MaxCentersReachedException(const char * m) : message{m} {}
+    virtual const char * what() const noexcept override {return message.c_str();}
+private:
+    std::string message = "";
+};
 
 template<typename dtype>
 class RegularSpaceClustering : public ClusteringBase<dtype> {
@@ -33,33 +45,36 @@ public:
      * @param chunk array shape(n, d)
      * @param py_centers python list containing found centers.
      */
-    void cluster(const py::array_t <dtype, py::array::c_style> &chunk, py::list& py_centers) {
+    void cluster(const py::array_t <dtype, py::array::c_style> &chunk, py::list& py_centers, unsigned int n_threads) {
         // this checks for ndim == 2
         const auto& data = chunk.template unchecked< 2 >();
 
         std::size_t N_frames = chunk.shape(0);
         std::size_t dim = chunk.shape(1);
         std::size_t N_centers = py_centers.size();
+        #if defined(USE_OPENMP)
+        omp_set_num_threads(n_threads);
+        #endif
         // do the clustering
         for (std::size_t i = 0; i < N_frames; ++i) {
-            dtype mindist = std::numeric_limits<dtype>::max();
-
+            auto mindist = std::numeric_limits<dtype>::max();
+            #pragma omp parallel for reduction(min:mindist)
             for (std::size_t j = 0; j < N_centers; ++j) {
                 // TODO avoid the cast in inner loop?
                 auto point = py_centers[j].cast < py::array_t < dtype >> ();
-                dtype d = parent_t::metric.get()->compute(&data(i, 0), point.data());
+                auto d = parent_t::metric.get()->compute(&data(i, 0), point.data());
                 if (d < mindist) mindist = d;
             }
             if (mindist > dmin) {
                 if (N_centers + 1 > max_clusters) {
-                    throw std::runtime_error(
+                    throw MaxCentersReachedException(
                             "Maximum number of cluster centers reached. Consider increasing max_clusters "
-                                    "or choose a larger minimum distance, dmin.");
+                            "or choose a larger minimum distance, dmin.");
                 }
                 // add newly found center
                 std::vector<size_t> shape = {1, dim};
                 py::array_t<dtype> new_center(shape, nullptr);
-                std::memcpy(new_center.mutable_data(), &data(i,0), sizeof(dtype)*dim);
+                std::memcpy(new_center.mutable_data(), &data(i, 0), sizeof(dtype)*dim);
 
                 py_centers.append(new_center);
                 N_centers++;